haskell.packages.ghc96.haskell-language-server: fix build
This dependency is only needed for the "plugin-tutorial" executable.
This is only built with GHC 9.6 and 9.8, but the expression is created
with GHC 9.10. Thus add the missing dependency.

It would be better to just disable that tutorial, but I didn't find a
quick way of doing that - toggling the related flag did not work.

This only fixed the build for GHC 9.6 for me, GHC 9.8 failed with a
different error right after.