various: remove `allowSubstitutes = false`
This commit removes explicit `allowSubstitutes = false` from 99% of usages
in-tree, excluding some special cases like test FODs that we can just
leave be for now, and impure derivations that need it for obvious
reasons.

Historically, this attribute has caused much more pain compared to any
compute time it'd save, so much so that Nix has a flag to completely
ignore it!