ci/github-script/bot: fix concurrency limit
This was introduced as part of the hotfix PR to avoid hitting API rate
limits - but the condition was wrong. It was supposed to trigger in all
PR contexts, not only for the Test workflow.

(cherry picked from commit a146035a2b1295b2d349fc1762c04cb4fb1e1f0b)