kernel: modules-closure: allowMissing can be an array of module names
I am building a minified kernel for an arm64 machine without autoModules
= true and which has LUKS enabled. LUKS would like to make cryptd
available to loaded in the initrd and does so by adding it to
boot.initrd.availableKernelModules. The default configuration for arm64
does not enable CONFIG_CRYPTD and the module is not built.
makeModuleClosure does not allow modules to be missing by default, thus
the shrunk modules closure fails to build. I can set
boot.initrd.allowMissingModules = true, but I'd still like to
audit/understand which modules are missing and either bless them to be
missing or ensure the module is built. Update "allowMissing" to be
either a bool or (new) a list of modules which can be missing.