rubyPackages: update, reverting regressions...
While this reverts regressions, this has to cause "new" regressions.

The problem here is that the previous regressions are relatively recent,
still, and the previous regressions are ***much worse*** than the slight
newer regressions that fixes the older regressions.

Note that this is not regressing the packages that are "newly"
regressing prior to the versions they were beforehand. They are now
currently pinned in a pessimistic manner, to correctly fix the problems
in the way that was initially intended.

Related PRs:

 - https://github.com/NixOS/nixpkgs/pull/442347
 - https://github.com/NixOS/nixpkgs/pull/443679