zoom-us: simplify updater script
This uses pkgsCross to help the general
update script find and update the various
zoom versions for the various architectures.
That way, we no longer need to calculate
the source hashes in the update script.

Credits:
The commit at hand is inspired by
https://github.com/NixOS/nixpkgs/blob/c656e00a21df6c42d6b57e9f8abf6cdf20e7b1ce/pkgs/by-name/ch/chataigne/update.sh#L24
(note: not merged into nixpkgs yet, so that link might disappear),
after a friendly note by axelkar
https://github.com/NixOS/nixpkgs/pull/429215#discussion_r2255114820
and a helpful hint by magistau
https://github.com/NixOS/nixpkgs/pull/429215#discussion_r2506957507
Thanks to both of you!