nixos/nix-channel: default to channels.nixos.org
Infra is interested in deprecating the redirect over the domain apex
eventually, since it is just a redirect to channels.nixos.org, which
we can more reliably redistribute with a CDN.

The nixos.org domain currently offers no IPv6 due to issues we've had
with our DNS hoster. (cf. nixos/infra#873)