actions/install-nix: restrict eval by default
We used to do this for some jobs, which *eval* untrusted code - but it
has been lost along the way. Since these jobs also have access to the
cachix secrets, an attacker could read these at eval time and poison the
cachix cache later on.

Currently, this would likely only allow to bypass other checks, but not
gain elevated privileges - because we recently remove the only job that
ran something cached on cachix with secrets available. Still, this is
serious and likely to lead to more mistakes in the future.