nut: configure the user and group nut should be built with
The udev rules which are generated during build are defaulting to the
`nobody` user which does not exist and should not be used (ref: [#55370](https://github.com/NixOS/nixpkgs/issues/55370))

This was also raised upstream [networkupstools/nut#3173](https://github.com/networkupstools/nut/issues/3173) and was informed that the `nobody` user will stay as is but they will consider adding a warning.
The user and group should be set by the builder (end user / distribution).