nixos/knot-resolver: init
We decided to split the NixOS modules for knot-resolver 5 and 6. For >=6
we also go forward with the (now standard) name knot-resolver.
This change also adds a warning to services.kresd if users try to use it
with a knot-resolver 6 package.

Co-Authored-By: Vladimír Čunát <v@cunat.cz>
(cherry picked from commit fd66108af069817c76061218f6791289a7607b75)