emacs-small: init at 30.2
Adds the "-small" variant of Emacs to nixpkgs to allow users to
install a distribution of GNU Emacs with the least amount of features
compiled in, essentially a "least means" GNU Emacs distribution. This
is done with an additional `.overideAttrs` call to manually override
the `nativeBuildInputs` and `buildInputs` attributes.

The motivation for this change is to introduce a variant of Emacs that
has a greatly reduced closure size, either for CI, portability and/or
minimalism.