arti: enable `full` feature flag
The `full` feature flag includes all stable and non-conflicting arti
feature flags. Upstream notes that the primary downsides are increased
memory usage during build and an increased binary size, both of which
are acceptable for the nixpkgs.

Some tests enabled by this feature flag do fail due to security checks
that are incompatible with the nixpkgs build sandbox, so those do need
to be disabled.