jmeter: fix HTML report generation
This commit fixes the HTML report generation feature (-g -o flags) that was
failing due to Nix store immutability. JMeter's report generator tries to
copy and modify template directories, which fails when files are read-only.

There is now a wrapper script that will handle this nix-specific case that jmeter wasn't designed to handle.