python3Packages.esp-idf-size: add 1.x argparse compat
Depending on the esp-idf version 1.x or 2.x can be required. The
difference is that 1.x provides the --ng flag, which 2.x does not.

The --ng behavior is the default in 2.x, so adding the flag as a stub
so the argparser does not complain is a cheap escape hatch to support
both.