nixos/file-roller: drop
Drops the `programs.file-roller` option as it is not needed for the
program to function and removes all instances of it being used within
nixpkgs.

In addition, file-roller is also removed from being included by
default in GNOME as it's not part of the GNOME core applications and
file archival functions are handled by nautilus (GNOME Files).