llama-cpp: remove git build dependency
`git` is used in upstream build system to determine the package version
from git history depth. We are not running from a git checkout, so `git`
is useless. (We handle version calculation during `src` preparation.)

Moreover, having `git` in the build scope results in versions captured
in `cmake` files as `unknown` / `dirty`.