librsvg: fix GDK-Pixbuf SVG loader on Darwin
The switch to Meson causes an issue on Darwin, where the GDK-Pixbuf loader is installed with a .dylib extension instead
of a .so extension.  This fixes the file extension, the @rpath-reference and the install name of the loader library.

see https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1049
see https://gitlab.gnome.org/GNOME/glib/-/issues/1413
see also #41559