nixos-rebuild-ng: use remote journalctl instead of pipe for logs
Notably, `--pipe` would cause `nixos-rebuild-ng` to panic when stderr
closes. That's especially painful when there's changes to the systemd unit
that manages the network connection, since that wouldn't come back up.

This materially improves the reliability of 'nixos-rebuild switch' and
while it adds a bit of complexity it doesn't seem too bad. NixOS test
coverage of nixos-rebuild is also pretty OK so that gives confidence
it won't break too much, except for use cases that rely on passing
information from the 'deployer' to the 'target' machine via stdin,
which seems like an obscure pattern.

Fixes #462179