bazel_8: add test for linux-sandbox spawn strategy
Upstream bazel is probing `/bin/true` to check if given sandbox strategy
works, it fails on NixOS and bazel will select weaker sandbox (or fail
the build if requested to use sandbox via `--spawn_strategy=linux-sandbox`)

The test is just taking `examples.cpp` and building it with
`--spawn_strategy=linux-sandbox` argument on Linux platforms.

https://github.com/NixOS/nixpkgs/pull/461754 has a patch to fix this,
verified that currently test fails and with a patch it passes.