google-compute-image: use systemd-repart for EFI boot...
...and enable systemd-boot in that case. Support for other bootloaders
can be added later on, if needed.

Support for legacy MBR booting (with GRUB) is retained, but emits a
warning that it will be removed before 25.06 is released unless it
turns out there is a reason it's still needed.

EFI booting has been the default on Google Compute Engine since
April 2020, 5.5 years ago.

Reasons for systemd-repart include not requiring a VM to build images
as well as more flexibility w.r.t. disk layouts compared to
`make-disk-image.nix`.