stdenv: make inputDerivation never require system features
This change removes a limitation with inputDerivations. See
stdenv.tests.inputDerivationRequiredSystemFeatures.meta.longDescription
for details.

I had originally ran into this limitation when trying to write a test
for pull request #319420. That pull request adds a setup hook that can
only be used with derivations that have "uid-range" in their
requiredSystemFeatures. I want to be able to build the inputDerivation
for part of the test on systems that don’t have that system feature.