unzip: inline patch files from dead sources
Of all the patches listed in unzip, these two stand out as depending on
dubious or unreliable sources. Concretely, the Internet Archive was down
yesterday (Nov 20, 2025) and caused build failures for the unzip
package. I also can't imagine that the IA folks want to be a hard
dependency for this.

There is prior art for inlining unzip patches into the repository and
these two are quite small.