buildFHSEnvBubblewrap: do not use read-only bind on /nix
Using a read-only bind mount on /nix for bubblewrap FHS environments
means that the build products of buildFHSEnvBubblewrap cannot write to
anywhere in /nix (notably /nix/var/nix/*, which is not a read-only
directory in any deployment I've seen). Since /nix/store is in most
cases already read-only on the host system (thanks to
`boot.nixStoreMountOpts` defaulting to options that make it read-only),
there is no need to enforce that property on the bubblewrap side as
well.