haskell: export build flags as a shell variable
This is useful so that you can call `Setup build` on other targets e.g.
for code generation while still having the same flags.

The context that led me to write this patch is that we have HIE file
generation in an overlay at work and I was packaging Glean, which
requires a code gen component get built, and it was building twice due
to the hie flags not getting propagated to the Setup build commands I
was running manually.