serverless: drop
Version 3.x of Serverless is no longer maintained by Serverless, and
has various security vulnerabilities.

Version 4.x is maintained upstream, but it is now closed-source.
We cannot simply switch the license to "unfree" because, although there
is a way to download the current release binary from Serverless,
it is no longer suitable for packaging in Nixpkgs.

Specifically:
* The only way to download a binary is unversioned, so builds of old
  versions would either fail, happen to work if it is cached, or would
  silently build the new version.
* Automatic updates are enabled, and the updates go to your home
  directory (and the updated binary either does nothing or silently
  replaces the packaged version in your PATH).

These could be worked around if it were possible to build from source,
but it isn't: see serverless/serverless#12761.

In light of all of this, I think this package should be removed.