reaper, libswell: fix directory opening
Replace bundled libSwell in reaper. The original one has a hardcoded
path to xdg-open. It's not the best solution, but upstream ignored the
provided patch for a long time and we can't really identify which
version of libswell is bundled.

Also, to make xdg-open work we need to ensure the right version of
openssl is available since it's loaded as a transitive dependency.