lemmy-ui: fix asset URLs
As `fetchFromGitHub` does not preserve the `.git` directory,
lemmy-ui fails to derive the value for its `COMMIT_HASH` variable
during the build.
The package previously (and mistakenly) assumed that `src.rev`
would somehow return the commit hash and not the version.
Fortunately, we do not need to know the precise commit rev.
This value is only used to generate the templated URLs without
further processing (and is also only injected for cache invalidation).