musl: fetch stdio patch to fix systemd procfs interactions
This fixes systemd oomd and user namespacing.
Previously musl would write an empty iovec into unbuffered procfs file
streams, resulting in the kernel rejecting the file write and systemd
throwing errors as a result. Alpine and Fedora already use this patch,
systemd lists this patch as expplicitly required for it to work on musl
following the official upstream musl support.