nixos/ifstate: remove netns so ifstate can identify interface
This test has never worked because when 'identify' is not configured,
ifstate searches for an interface with the specified name inside the
defined network namespace. However, virtualisation.interfaces cannot
place an interface into a specific network namespace, and without
identifying details (such as a MAC address or bus ID), ifstate is
unable to locate the correct interface.

Removing the network namespace avoids this problem.