[WIP] nixos/systemd-confinement: Fix template units
This is to reproduce the regression that has been reported at
https://github.com/NixOS/nixpkgs/issues/464323.

The reason why I didn't use the same implementation that the issue
reporter is using is because they're relying on sleep and the last thing
we want is more flaky tests.

A possible fix for the issue would be to use "/run/confinement/%n" as
the root directory instead of "/run/confinement/${mkPathSafeName name}"
and using the implementation from the issue reporter it indeed seems to
fix the issue.

What I'm currently not sure is *why* this issue happens exactly, apart
from "well there is some conflict because both instances share their
root directory".

I already diffed the outputs form systemd.log_level=debug and apart from
missing directories in the service that has started first, there were no
differences.

TODOs:

  * Obviously, we want to make sure the test fails without the fix.
  * Run nixfmt.

Signed-off-by: aszlig <aszlig@nix.build>