nixos/wpa_supplicant: ignore shellcheck nitpick
This can occur if one enables `systemd.enableStrictShellChecks`.

Shellcheck itself writes on this notice that one should use arrays in
this situation. I tried to do that, but the curly braces with nix
variable expansion gets boring very quickly. As such, we just ignore
this nitpick, as the use is reasonable here.
https://www.shellcheck.net/wiki/SC2086

Also enable `enableStrictShellChecks` to avoid regressing.