jellyfin-media-player: 1.12.0 -> ???
WIP, waiting for upstream to cut a new release.

Creating this patch early on to benefit from the newly acquired
upstream qt6 support.

I got rid of the web patch. It seems to be outdated.

TODO: review all the libs this derivation is depending on. I have the
feeling we probably can get of some.