yarn2nix-moretea: update dependencies for yarn2nix
There are a lot of major version updates here, but I've tested this on itself,
and the updated version produces identical output to 25.05's version.

I also ran eslint and the formatter in this commit due to the bumps to both of those packages.

This fixes all security issues with dependencies, including CVE-2022-25860,
CVE-2020-7707, CVE-2022-25912, and about 10 others.