boulder: 2025-04-17 -> v0.20251118.0, fix build
Boulder has mostly removed keys and certificates for testing in the
meantime. They have to be generated on-the-fly. The ipki keys just
require minica. The webpki keys would require softhsm and this somehow
fails. We currently get away without them, probably because we have
disabled all the tests that require them.