nfs-utils: remove unused tcp-wrappers dependency
nfs-utils _can_ be built with tcp-wrappers support, but this is disabled
by default and needs a `--with-tcp-wrappers` configure flag to be
enabled.

Since we never noticed this was not supported, and tcp-wrappers is
unmaintained in nixpkgs (as is nfs-utils), I think we should just not
include this support for now.