tamarin-prover: add patch to allow maude 3.5.1
This patch fixes the build failure of tamarin-prover.
The latest version of maude was not allowd in tamarin, which was causing
the build failure.
This patch is in upstream, so it should be removed on the next release.

(cherry picked from commit 1ef3d3a127cc48fd2995d442d9aef31791583f98)