nix: Fix splicing
This is a port of
https://github.com/nixos/nix/commit/cb5b0c30aa5733b40cc70089fe3829cb1046c352
from https://github.com/NixOS/nix/pull/14489 but adopted to preserve
dependency order, and this not cause rebuilds.

We only need the one commit of that PR because we don't care about a dev
shell for Nix in Nixpkgs.

(cherry picked from commit 8f72b28d567f0e576c107fe5b80acdd6e6961806)