lib.strings.escapePythonBytes: avoid many unnecessary escape sequences
See the comment above escapePythonBytes for details.

When making this change, I had considered going farther. I had
considered making it so that lib.strings.escapePythonBytes uses as few
escape sequences as possible. I decided that doing that would be overly
hard and error-prone. If you want to use as few escape sequences as
possible, then you need to choose which of the following types of bytes
literals would result in the least number of escape sequences:

• b'Example'
• b"Example"
• b'''Example'''
• b"""Example"""
• br'Example'
• br"Example"
• br'''Example'''
• br"""Example"""

I decided that writing code that would make that choice correctly was
not worth the effort.