nixos/ddclient: fix missing 'password=' config key
The `password=` prefix was missing for the `protocol == nsupdate` and
`secretsFile != null` cases. As a result, there's a line in the
generated config file containing only the path to the secrets file / the
`secrets_placeholder`, which `ddclient` interprets as a hostname for
which to update a DNS entry, which won't work for obvious reasons.

(cherry picked from commit 1af18e36bfdd7364e81a2e31e1807b875bcb4202)