firefox: don't touch dontFixup on darwin if it's false
To fix the wrapper for firefox-bin, we now force dontFixup == false on
Darwin: #457408

The change unnecessarily touched the attribute on a package when it's
not set, making `firefox` wrapper build its own firefox-unwrapped (with
the attribute forcefully set to false).

In addition to unnecessary duplication of nix outputs, since `firefox`
is not built in Hydra, it means that all users are now forced to build
firefox (and thunderbird) when using the wrapper.

This patch fixes the issue by forcing the attribute only when it's
present and set to true in the wrapped package.