gnomeExtensions.forge: package manually
This extension has not gotten a new release in while, and the latest one
in the GNOME extension repository is not compatible with GNOME 49. This
commit replaces that one with a new one manually packaged directly from
GitHub, and which includes fixes to make it work with GNOME 49 again.

(cherry picked from commit 1db54da55a20a3d35526802dbff72b817d03b905)