ovn: fix command presence detection during checkPhase
The test suite uses `which` to detect if auxiliary dependencies like
scapy or tcpdump are installed on the system, and skips the tests using
these commands if not.

Problem is when `which` is missing, all such commands are detected as
missing, so some tests that could be safely run are unnecessarily
skipped.

Note: upstream should probably not swallow the error from missing
`which` and instead enforce its presence during build.

The patch doing the above is posted at:
https://patchwork.ozlabs.org/project/ovn/patch/20251207205313.27750-1-ihar.hrachyshka@gmail.com/